LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/document-exit-fullscreen-active-document-expected.txt


PASS Document#exitFullscreen() when the document is not the active document
